# backend/src/main/java/com/ticketmanagement/ticket/domain/TicketStatus.java
package com.ticketmanagement.ticket.domain;

public enum TicketStatus {
  OPEN("Open"),
  IN_PROGRESS("In Progress"),
  RESOLVED("Resolved"),
  CLOSED("Closed");

  private final String label;

  TicketStatus(String label) {
    this.label = label;
  }

  public String label() {
    return label;
  }
}